Add base option to `phylum analyze`
This patch adds the new `--base` option which allows passing a file
which contains a list of packages that will be ignored for analysis.
This can be used to analyze the delta between two package lists.

The expected file format is identical to the output of `phylum parse`,
so a base can be obtained using `phylum parse > base.json` and then
analyzed using `phylum analyze --base base.json`.

The option is hidden and not intended to be used by endusers, mainly to
avoid having to document the exact required file format.

Closes #1003.