Skip analysis with empty package list (#1007)
When sending an analysis request with an empty list of packages, the API
will return an error. Since we know ahead of time that this will be the
case, it shouldn't be necessary to actually make the request.

This patchs adds a simple warning message telling the user clearly what
the issue is and then exits with a zero exit code before any requests
are made. This also reduces the noise in the output making it easier for
users to understand the issue.