Reformat `phylum history` output
The format of the `phylum history` command was very hard to read and
littered with unlabeled or useless information. This patch brings a new
simplified format.

The new format avoids long lines, but it does so by using more lines
(currently 8 lines per job). This is a necessary trade-off. I first
tried a table format, but the lines were 150+ characters long.