Remove request type from global config (#1001)
* Make `phylum batch` type argument mandatory

This change accompanies 421fa11 which made the type argument mandatory
for the `phylum package` command.

Using the package type from global configuration file is confusing, so
this change removes that confusion by making the argument mandatory.

* Remove package type from global configuration file

We are no longer using this value, so it can be removed.

_Note_: Configuration files written by the CLI after this change will
not be readable by any CLI versions before this change. So once a user
has upgraded beyond this change, they cannot go back.

* Use clap .value_parser for `phylum batch --type`

The raw string here was also causing rustfmt to ignore this code. So I
have split this change into two commits: The basic code change (this
commit) and the rustfmt

* Reformat app.rs

Now that this code block no longer includes a raw string, rustfmt has
some changes to make. This commit is only a reformat.

* Strengthen config::write_parses_identical test

The previous test only checked a single value. This patch updates the
test to check the entire structure.