Reformat `phylum history` output (#1010)
* Reformat `phylum history` output

The format of the `phylum history` command was very hard to read and
littered with unlabeled or useless information. This patch brings a new
simplified format.

The new format avoids long lines, but it does so by using more lines
(currently 8 lines per job). This is a necessary trade-off. I first
tried a table format, but the lines were 150+ characters long.

* Use the same language as the report for status

The policy report uses specific language for the pass/fail/incomplete
cases. This patch updates our language to match the report.