ci: switch `ubuntu-22.04` to `ubuntu-latest` (#1023)
`ubuntu-latest` is now the same as `ubuntu-22.04`. The specific version was previously used to ensure a newer version of `shellcheck` was installed and available on the runner, so that a particular command option would be available. `ubuntu-latest`, even if it progresses, should still have at least that version of `shellcheck` installed.

Closes #467