Add generator for manifest to lockfile conversion
This patch adds the new `lockfile_generator` crate to the project, which
takes in a project directory and attempts to non-mutably generate a
lockfile for it.

This crate is used by the CLI's lockfile parser to automatically
generate a lockfile if the lockfile path it has received is a manifest
file rather than a lockfile.

The type of analysis required is determined in the following steps.

1) Format determination:
 - Try to determine lockfile type from CLI arguments
 - Try to determine lockfile type from lockfile name
 - Try to determine lockfile type from manifest name:
    - Find a format for the manifest that already has a lockfile
    - Fallback to first format that matches the manifest

2) Lockfile analysis:
 - If 1) did not hit the manifest fallback path, we analyze the lockfile
   with the detected format's parser.

3) Lockfile generation:
 - Check if original non-lockfile path is a manifest
 - Check if there is a generator available for the format
 - Generate lockfile at manifest's location using the format's generator
 - Analyze the generated lockfile

The split in analysis -> manifest determination -> analysis is necessary
since some formats have files that are commonly used as both lockfiles
and manifests (requirements.txt), thus it is necessary to first attempt
a lockfile analysis before being able to re-attempt parsing the
generated lockfile.

Closes #1046.