fix: `pip` parser fails for some lines containing comments
The `pip` "Requirements File Format" is documented here:
https://pip.pypa.io/en/stable/reference/requirements-file-format/

It allows for line continuations, which this change does not address.
Instead, this change accounts for lines containing comments where the
comment character (`#`) is not the first character of the line. The
comment line could be indented or comments can be inline (e.g., after
a valid requirement). Both types of comments were added to the test
fixture.

When updating the `CHANGELOG.md`, saving the file in the IDE caused an
automatic reformatting of some lines...to remove trailing whitespace and
ensure consistency for the list markers (`-` everywhere instead of `*`).

Closes #1044