refactor: use `tag_name` instead of `name` to identify releases
The `name` entry stores the GitHub Release name, which could be set to something other than the version. Using the `tag_name` entry is better since the tags are much more tightly coupled with the release version. A check of the 30 most recent CLI releases showed that the `name` and `tag_name` values were identical for every release. This means there is no worry that this change will break updates from older releases to newer ones.