Add manifests to recursive lookup
This patch adds manifest files to the files returned by the recursive
lockfile lookup code. This affects both `phylum init` and analyze/parse
without project file or CLI override.

Manifest files are filtered out when a lockfile is present in a parent
directory of that manifest. The assumption is made that lockfiles in
directories above manifests cover all manifests below them, or there's
additional lockfiles for those subdirectories.

Closes #1048.