Fix parsing with empty yarn lockfile
This patch resolves an issue in both the v1 and v2 yarn lockfile parsers
where a newly created lockfile would cause the parser to throw an error.

With yarn v1 the generated lockfile will only contain the lockfile
header, thus parsing the file and checking for emptiness is sufficient
to detect the edgecase.

With yarn v2 we cannot actually detect that a v2 yarn lockfile is
present, since the lockfile is completely empty and does not contain any
version info. However by making the version header optional in the v1
parser, it's still possible to parse the empty v2 lockfile using the v1
parser fallback.

Closes #1028.