docs: update `yarn` extension usage
The command used in the `yarn` extension docs for usage was not correct.
It specified `yarn install my-package`, which results in an error:

`Unknown Syntax Error: Extraneous positional argument ("<my-package>")`

Indeed, the docs for [yarn install](https://yarnpkg.com/cli/install) do
not indicate a package to be specified on the command line. The correct
command is `yarn add`, which does take a package on the command line.