docs: update manifest generation list and language (#1090)
Some manifests were listed incorrectly. Some were missing. They have been updated to the current state. The language around the manifest generation was also updated to make it more clear that the tools used to perform the lockfile generation are required to be installed and available in the environment where the CLI is run. Finally, the list of lockfiles was updated to follow the same sequence as the manifest files. The output from the `phylum analyze` command on this document is still wrong but changing it now was purposefully eschewed.