Bump deno
Deno was updated in denoland/deno@1d447cb7 so that disabling an op
removes the underlying function.

For us, this means that Deno.permissions.request (and requestSync) will
fail instantly rather than acting as a no-op. That is likely beneficial
to alert extension authors that the function is not supported in Phylum
extensions.

Also, objects from our API now show "[Object: null prototype]" when
printed with `console.log()`. This is good, but it breaks our fragile
tests...

Using `JSON.stringify()` isn't much better... But it should be good
enough for now.