Fix status on PhylumApi.createProject()
The `status` field returned by `createProject()` has always either been
"Created" or "Exists" (since it was introduced in 488eaef). But the
test and docs showed different values. And this wasn't caught by the
test because of #1108...

This patch updates the test and the type system to reflect the correct
values.