Disable lockfile generation for yarn v1
Since yarn v1 isn't able to generate a lockfile without installing the
dependencies and executing their build code, it could be dangerous to
automatically generate lockfiles for the user.

To fix this issue, the lockfile generator will now return an error when
lockfile generation is attempted with the yarn generator and only
version 1 is used by the project.

While the lockfile generation with yarn v1 still worked and it would
provide protection against vulnerable code in the target executable, it
is safer to leave the lockfile generation up to the user so they're in
full control.

Closes #1086.