Filter setup.py with pyproject.toml present
This patch changes the behavior for loading Python manifest files by
ignoring `setup.py` files when a `pyproject.toml` is present at or above
the `setup.py`'s directory.

This is in line with what pip will do by itself when loading manifest
files, ensuring that our behavior is consistent with pip and we do not
unnecessarily generate the same lockfile twice.

Closes #1102.