Ensure correct cargo is used
This patch fixes an issue that was occurring on our self-hosted runner.
Since this runner also builds Homebrew bottles, it has `cargo` installed
via Homebrew.

Unfortunately, this can lead to the wrong `cargo` being used, which
means that we are not getting the version that we expect in our stable
test and our oldstable tests fail outright because the +<toolchain>
syntax does not work with `cargo` from Homebrew.

To ensure that we are using `cargo` from `rustup`, this patch explicitly
calls `rustup run`.