Add support for generating C# lockfiles
This patch adds a new C# lockfile generator using the `dotnet` tool,
which allows generating lockfiles for version .NET 5 and above.

The lockfile support has also been expanded to support
`packages.<project_name>.lock.json` in addition to `packages.lock.json`
since this was mentioned as an option in Microsoft's documentation.

This patch removes support for `*.csproj` files as separate lockfiles,
so this patch will remove support for projects below .NET 5. However
this seems like a reasonable tradeoff considering we do not have a big
C# userbase and this significantly streamlines and simplifies the C#
ecosystem.

The old csproj tests were removed in favor of generating a C# lockfile
from a project with a manifest. This requires developers and CI to have
the `dotnet` tool installed, but that should not cause any issues.

Closes #1142.