Reverse search direction for manifests' lockfiles
Previously when a manifest was specified as the target for an analyze or
parse operation, the directory would be walked downwards depth-first to
find a lockfile for this manifest. This would both prefer lockfiles
farther away from the manifest and allow lockfiles from subdirectories.

This patch reverses the search direction for lockfiles, since they're
generally at or above their corresponding manifest. The new logic
prefers lockfiles closest to the manifest and searches the directory
tree upwards from there.

Closes #1170.