Fix reported lines in extension errors
This fixes an issue with extensions where the reported line for an error
did not match the supplied code, since the error was based on the
transpiled JavaScript.

To make sure the error can be mapped from JavaScript back to the
original TypeScript, a transpiler cache has been introduced which is
used to retrieve the original source map when an error is encountered.

Fixes #961.