Use absolute paths for install check
This patch updates the install check to correctly handle relative paths.
Before this, it was possible to fool the check by using a relative path
on the command line.

```
> phylum extension install -y ~/.local/share/phylum/extensions/npm
Installing extension npm...
❗ Error: extension path and installation path are identical, skipping

> phylum extension install -y ./.local/share/phylum/extensions/npm
Installing extension npm...
❗ Error: IO error for operation on ./.local/share/phylum/extensions/npm: No such file or directory (os error 2)
```

I believe the previous check could also be fooled with an absolute path
on the command line if `XDG_DATA_HOME` contained a relative path. But
that scenario is very unlikely, so I did not test it.