Enable Web Storage API for extensions
This patch allows extensions to use `localStorage` to store state. Each
extension has separate state and it is removed when the extension is
uninstalled. Additionally, only installed extensions are allowed to
persist state.

For an example usage, consider this extension:

```js
function runCount() {
  try {
    const c = parseInt(localStorage.getItem("runCount") ?? "0", 10);
    localStorage.setItem("runCount", `${c+1}`);
    return c;
  } catch(_e) {
    // Local Storage Disabled
    return -1;
  }
}

console.log(`This extension has been run ${runCount()} times.`);
```

_Note:_ This also enables usage of `sessionStorage`, but the session
lasts only as long as the process, so this does not provide any
additional value.