Add extension API for Linux sandbox control
This patch adds two new fields to the `runSandboxed` extension API which
allow controlling the minimum required Linux landlock ABI version and
whether the command should still be executed when sandboxing is not
possible.

With the `best_effort` flag set, this also allows running extensions on
Windows now with the sandbox disabled when this flag is set.