docs: format `clap_markdown` output
This change ensures the subcommands are sorted alphabetically when shown in a list of the parent command. It also adds backtick formatting to all the option and value name representations. That way, the output will be formatted more clearly when rendered as markdown. The docs were re-generated with these changes.