Fix cargo lockfile generation (#1233)
When generating a lockfile for a Cargo manifest, it is possible that the
lockfile will be placed in a parent directory rather than the manifest's
directory. Since this wasn't accounted for, lockfile generation would
fail and the lockfile would not get cleaned up.

To allow tracking lockfiles generated in directories above the manifest,
the `locate-project` subcommand of cargo is used now to locate the place
where the lockfile will be put.