Fix requirements parsing with index URL option (#1251)
When using `pip-compile` with a third party Python index, the `--index-url`
and `--extra-index-url` options get included in the `requirements.txt` output
file. This patch updates our parser to allow these options.

The presence of `--index-url` in the file will make the parser assume that all
packages are being pulled from the third party registry. This may or may not
be true, but it doesn't really matter since we currently treat third party
registries as if they were first party when submitting a job.

Co-authored-by: Kyle Willmon <kyle@phylum.io>