Add find-lockable-files subcommand
This patch adds a new subcommand which allows looking up all lockfiles
and manifests in the current directory without performing any filtering
on the result.

The manifest resolution is also changed to remove manifests when there
is a manifest of the same ecosystem in a directory above, indicating
that the project is likely a workspace.

Closes #1255.