Switch to dependency file terminology
This patch makes sweeping breaking changes to change the internal and
public terminology to refer to a manifest or lockfile as "dependency
file".

This change should also clarify some CLI interfaces where the previous
terminology was asking for a lockfile, but a manifest file was accepted
without clearly stating it.

Closes #1271.