Move generic depfile parsing to phylum_lockfile
This patch moves a large chunk of the dependency file parsing from the
CLI into `phylum_lockfile` to allow others to make use of it.

The interface of the functions has been reworked based on the assumption
that filesystem access might not be present, allowing pure in-memory for
API usage.