Switch documentation hosting from Readme.com to Docusaurus
This is part of a larger effort to move <https://docs.phylum.io> from
Readme.com to Docusaurus, as a statically generated site. The new site
will host all the documentation, regardless of where the docs originate.
The CLI docs from this repository are included in the
`phylum-dev/documentation` repository as a git submodule. The format of
the markdown docs has been changed to account for operating in the new
structure and with Docusaurus as the static site generator.

Actions taken include:

* Update doc templates to remove unused Readme.com front matter
* Remove all doc template files that only existed to specify a category
  * The sub-categories are removed in favor of simpler command updates
* Update `gendocs` xtask
  * Remove Readme.com front matter in the default header
  * Regenerate all docs with `cargo xtask gendocs`