Parse manifest files with non-standard names (#1337)
This change allows for parsing manifest files with names that don't
match the expected standards. For example, `pip` manifest files can be
named *anything* and they are therefore hard to identify. Pip lockfiles
can also be named anything but they will at least get picked up by the
fallback approach of attempting to parse with all known parsers
(assuming a "true" lockfile with no loose requirement entries).

A new custom error (`UnknownManifestFormat`) and exit code (21) were
created to be used when a dependency file is provided without a file
type and does not parse as a known lockfile, making it a likely manifest
file.

A change was made to report errors with the `pip` version check that
happens in the `pip` lockfile generator.

Additionally, a change was made to ensure `ManifestWithoutGeneration`
errors use the display path instead of the full path.

Prerequisite checks were also added for lockfile generation to account
for known manifest file names. Previously, some lockfile generators
completely ignored the filename provided and just called the relevant
tool in the directory, leading to some unexpected successes.