Allow execution for apt-installed `maven`
When Maven is installed on Debian-based systems via `apt`, it gets
installed in `/usr/share/maven`, which is not included in the current
sandbox exceptions. This patch allows execute so that it will work...
hopefully.

If it still doesn't, there is a procedure for determining the complete
set of exceptions needed.