ci: use GitHub-hosted macOS runners (#1470)
The GitHub-hosted macOS runners have been updated to include arm64 M1 options now. This PR switches our workflows to use them instead of the self-hosted runner. There is more availability for these runners and they run just about as fast, if not faster when accounting for parallel jobs. The `release` workflow was updated to change `macos-latest` to `macos-14` to be specific about the version in the matrix and allow for adding additional versions in the future.

Resources:
* [Standard GitHub-hosted runners for public repos](https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners/about-github-hosted-runners#standard-github-hosted-runners-for-public-repositories)
* [Latest `macos-14` image release](https://github.com/actions/runner-images/releases/tag/macos-14-arm64%2F20240701.9)
* [SW included in latest `macos-14` image](https://github.com/actions/runner-images/blob/macos-14-arm64/20240701.9/images/macos/macos-14-arm64-Readme.md)