Add org support to `phylum group` (#1488)
This patch adds org support to the `list`/`create`/`delete` subcommands
of `phylum group`. The `member` subcommand is unchanged, since user
management is performed on the org level instead.

The `create` and `delete` subcommands will always use the currently
linked organization, with no way to temporarily override it. While this
could be annoying to some, it avoids adding extra flags that would get
deprecated in the future. Most users likely always or never work within
organizations and for people working in-between it's easiest to avoid
linking orgs and just using `--org` instead as required.

See #1482.