Update `yarn` extension sandbox exception (#1495)
This change allows for local project directory write access
for the building stage of the `yarn` extension. This brings
it into parity with the rest of the package manager extensions.

The need for this change was discovered by a community edition
user who reported in Discord how the extension was failing due
to a sandbox access violation:

```
Error: EROFS: read-only file system, rmdir '/home/sum/DEV/ETH/YD/yd-marketplace/node_modules/.store'
```

The same user also noted their confusion over the language
provided by the extension after this failure:

> Please submit your lockfile to Phylum should this error persist.

They did not have a lockfile present (just a `package.json` manifest)
and noted that the extension did not create a `yarn.lock` lockfile
for them to submit.

Therefore, the language of this message has been updated to be less
specific about lockfiles and more about "dependency files."

Finally, a troubleshooting section was added to the package manager
extensions' documentation. This is an attempt to better prepare users
for what to do when the extension fails.