Apply PR changes
The signed Windows binary change has been reverted.

The test rollup job has been reverted so that it now only contains those
jobs containing a matrix since the names are long and or changing over
time, making it hard to enforce them in branch protection rules of the
repo settings.

The `clippy` and `oldstable` jobs were updated to include the Windows
runners in their matrices.