Allow whitespace in `pip` requirement specifiers
A bug was discovered recently when attempting to parse a `pip` lockfile.
That is, a `requirements*.txt` file containing exact versions
(i.e., `==`) in all of it's requirement specifiers. Such a file should
parse as a lockfile and not fall back to lockfile generation. However,
if whitespace is included around the `==`, the parsing fails.

Whitespace is allowed in requirement specifiers. This change accounts
for that. It also updates the test fixture to prevent regressions.

Ref: https://pip.pypa.io/en/stable/reference/requirement-specifiers/