Add C# *.packages.config support
This adds support for `packages.config` files for the C# ecosystem under
the new `NugetConfig` lockfile name.

The file format itself seems to be relatively simple, with just a list
of package names and optional versions. The current test fixture is
based on examples of this lockfile found on <https://github.com>.

Both `packages.config` on its own and `packages.*.config` are supported
as valid names. Since this seems to be a lockfile there is no need for
lockfile generation.