Add `phylum firewall log` subcommand (#1551)
This patch adds the new `phylum firewall log` to allow checking the
Aviary log from the CLI.

While the UI gives a more user-friendly overview, this should be
complementary by giving access to the data in a structure closer to the
original layout.

The CLI arguments allow applying most filters exposed by the API, but
combine the individual package components into a PURL to avoid an
excessive number of arguments.

Currently neither the `--json` output nor the pretty-printer have a way
to paginate through the logs. While this could be done either
automatically or by manually specifying the pagination offset, it seems
to me like the limit of 10_000 entries per page should be sufficient for
all current usecases.

Contrary to our existing timestamp columns, the timestamp in the output
table uses nanosecond precision. The additional precision seems
worthwhile in this case since this automatically groups versions
together if they were submitted for analysis as a batch.