Fix Windows Gradle manifests incorrectly retained (#1553)
This change fixes a bug that is specific to Windows systems. When legacy
Gradle (prior to v7.0.0) lockfiles exist (e.g.,
`gradle/dependency-locks/*.lockfile`) along with a manifest at a higher
level, the manifest is meant to be ignored when finding dependency files
to analyze. However, on Windows systems this was not the case before
this change. Instead, both the manifest and the lockfiles were retained.
After this change, only the lockfiles are retained when both are
present.

A failing test was added first to confirm the incorrect behavior. Then,
a change was made to handle paths as `Path` objects instead of strings.
That way, the generic `/` path separator character will be used
correctly regardless of the runtime system.

Additionally, an error message was updated to reflect the correct
dependency file language for Gradle parsing errors.