Update workflow runners (#1590)
This change updates the workflow runners to move away from the soon to
be deprecated `ubuntu-20.04` image and use `ubuntu-22.04` instead. The
intention is to continue to use the oldest supported version to ensure
binaries can build and run on that version. `ubuntu-latest` is used for
the linting related jobs.

Additionally, uses of `apt` were changed to `apt-get` to silence
warnings in CI about how `apt` is not meant to be used in scripts.

Ref: https://github.com/actions/runner-images/issues/11101