Add `Unreleased` section back to `CHANGELOG.md` (#1601)
This change fixes the `release` workflow. It has been broken since the
v7.5.0 release, when the `CHANGELOG` was updated by hand instead of with
the `bump_version.sh` script. That is, the `Unreleased` section was
removed from the `CHANGELOG`, which caused the `Create release artifacts`
step of the `release` workflow to fail because it looks for the
`Unreleased` header and does not find it.